Conflicts:
	apps/pollster/models.py
	settings.py

+++ apps/pollster/models.py
@@ -4,6 +4,7 @@ from django.db import models, connection, transaction, IntegrityError, DatabaseE
 from django.contrib.auth.models import User
 from django.forms import ModelForm
 from django.core.validators import RegexValidator
+from django.template import Template
 from cms.models import CMSPlugin
 from xml.etree import ElementTree
 from math import pi,cos,sin,log,exp,atan
@@ -14,6 +15,8 @@ from django.conf import settings
 
 DEG_TO_RAD = pi/180
 RAD_TO_DEG = 180/pi
+
+import mapnik2
 
 try:
     import mapnik2 as mapnik
@@ -1005,7 +1008,9 @@ class Chart(models.Model):
     status = models.CharField(max_length=255, default='DRAFT', choices=CHART_STATUS_CHOICES)
     geotable = models.CharField(max_length=255, default='pollster_zip_codes', choices=settings.GEOMETRY_TABLES)
     hastitle = models.BooleanField(default=False)
-    
+    template = models.TextField(blank=True, default='')
+    realtime = models.BooleanField(default=False)
+
     class Meta:
         ordering = ['survey', 'shortname']
         unique_together = ('survey', 'shortname')
@@ -1026,6 +1031,10 @@ class Chart(models.Model):
         return self.status == 'PUBLISHED'
 
     @property
+    def is_template(self):
+        return self.type.shortname == 'template'
+
+    @property
     def has_data(self):
         if not self.sqlsource:
             return False
@@ -1043,7 +1052,7 @@ class Chart(models.Model):
             rows = [{"c": [{"v": v} for v in c]} for c in cells]
             data["dataTable"] = { "cols": cols, "rows": rows }
 
-        else:
+        elif self.type.shortname[:10] == "google-map":
             if self.chartwrapper:
                 data["bounds"] = json.loads(self.chartwrapper)
             try:
@@ -1220,27 +1229,20 @@ class Chart(models.Model):
                                   FROM %s B, (SELECT * FROM %s) A
                                  WHERE upper(A.zip_code_key) = upper(B.zip_code_key)""" % (geo_table, table,)
             cursor = connection.cursor()
-            #try:
             cursor.execute("DROP VIEW IF EXISTS %s" % (view,))
             cursor.execute("DROP TABLE IF EXISTS %s" % (table,))
-            cursor.execute("CREATE TABLE %s AS %s" % (table, table_query))
-            if hasattr(settings, 'POLLSTER_GRANT_CHART'):
-                for g in settings.POLLSTER_GRANT_CHART:
-                    cursor.execute("GRANT SELECT ON TABLE %s TO \"%s\"" % (table, g))
-            if self.type.shortname != 'google-charts':
-                cursor.execute("CREATE VIEW %s AS %s" % (view, view_query))
-            transaction.commit_unless_managed()
-            self.clear_map_tile_cache()
+            if not self.realtime:
+                cursor.execute("CREATE TABLE %s AS %s" % (table, table_query))
+                if self.type.shortname[:10] == "google-map":
+                    cursor.execute("CREATE VIEW %s AS %s" % (view, view_query))
+                transaction.commit_unless_managed()
+                self.clear_map_tile_cache()
             return True
-            #except IntegrityError:
-            #    return False
-            #except DatabaseError:
-            #    return False
         return False
 
     def update_data(self):
         table_query = self.sqlsource
-        if table_query:
+        if table_query and not self.realtime:
             table = self.get_table_name()
             cursor = connection.cursor()
             cursor.execute("DELETE FROM %s" % (table,))
@@ -1251,8 +1253,12 @@ class Chart(models.Model):
         return False
 
     def load_data(self, user_id, global_id):
-        table = self.get_table_name()
-        query = "SELECT * FROM %s" % (table,)
+        if not self.sqlsource:
+            return ((('Error',),), (("SQL query is missing",),))
+        if self.realtime:
+            query = "SELECT * FROM (%s) A" % (self.sqlsource,)
+        else:
+            query = "SELECT * FROM %s" % (self.get_table_name(),)
         if self.sqlfilter == 'USER' :
             query += """ WHERE "user" = %(user_id)s"""
         elif self.sqlfilter == 'PERSON':
@@ -1260,15 +1266,22 @@ class Chart(models.Model):
         params = { 'user_id': user_id, 'global_id': global_id }
         query = convert_query_paramstyle(connection, query, params)
         try:
+            sid = transaction.savepoint()
             cursor = connection.cursor()
             cursor.execute(query, params)
+            transaction.savepoint_commit(sid)
             return (cursor.description, cursor.fetchall())
         except DatabaseError, e:
+            transaction.savepoint_rollback(sid)
             return ((('Error',),), ((str(e),),))
 
     def load_colors(self, user_id, global_id):
-        table = self.get_table_name()
-        query = """SELECT DISTINCT color FROM %s""" % (table,)
+        if not self.sqlsource:
+            return ((('Error',),), (("SQL query is missing",),))
+        if self.realtime:
+            query = "SELECT DISTINCT color FROM (%s) A" % (self.sqlsource,)
+        else:
+            query = "SELECT DISTINCT color FROM %s" % (self.get_table_name(),)
         if self.sqlfilter == 'USER' :
             query += """ WHERE "user" = %(user_id)s"""
         elif self.sqlfilter == 'PERSON':
@@ -1276,10 +1289,13 @@ class Chart(models.Model):
         params = { 'user_id': user_id, 'global_id': global_id }
         query = convert_query_paramstyle(connection, query, params)
         try:
+            sid = transaction.savepoint()
             cursor = connection.cursor()
             cursor.execute(query, params)
+            transaction.savepoint_commit(sid)
             return [x[0] for x in cursor.fetchall()]
         except DatabaseError, e:
+            transaction.savepoint_rollback(sid)
             # If the SQL query is wrong we just return 'red'. We don't try to pop
             # up a warning because this probably is an async Javascript call: the
             # query error should be shown by the map editor.
@@ -1321,6 +1337,26 @@ class Chart(models.Model):
         except DatabaseError, e:
             return {}
 
+    def get_template(self):
+        return Template(self.template or "{% for row in rows %}{{ row }}<br/>{% endfor %}")
+
+    def render(self, context):
+        """Adds data to context and use it to render template."""
+        if self.type.shortname == "template":
+            template = self.get_template()
+            if template:
+                user_id = context["user_id"]
+                global_id = context["global_id"]
+                descriptions, cells = self.load_data(user_id, global_id)
+                cols = [desc[0] for desc in descriptions]
+                rows = [dict(zip(cols, cell)) for cell in cells]
+                context.update({"cols": cols, "rows": rows})
+                result = template.render(context)
+                context.pop()
+            else:
+                result = "Template is empty."
+            return result
+
 class GoogleProjection:
     def __init__(self, levels=25):
         self.Bc = []
@@ -1352,3 +1388,16 @@ class GoogleProjection:
 
 class SurveyChartPlugin(CMSPlugin):
     chart = models.ForeignKey(Chart)
+    show_on_success = models.BooleanField(default=False, verbose_name="Show on submit", help_text="Show this chart only on successful submit of its survey.")
+
+class SurveyPlugin(CMSPlugin):
+    survey = models.ForeignKey(Survey, limit_choices_to={"status":"PUBLISHED"}, verbose_name="Survey")
+    redirect_path = models.CharField(max_length=4096, blank=True, default='', verbose_name="Redirect to path")
+    success_template = models.TextField(blank=True, default='', verbose_name="Success template")
+
+    def get_template(self):
+        return Template(self.success_template)
+
+    def render(self, context):
+        return self.get_template().render(context)
+
